Split config handling into two cases
There are two cases we need to handle here:

(1) If we're given an explicit value for `--config`, then we should
error if it doesn't exist.

(2) If we're not given a value (the default is ""), then we need to look
for a `.vale.ini`. However, we can't send an error if we don't find one
because the user may simply be editing a non-Vale project/file.